requirements.txt
Django==4.0.5
psycopg2-binary==2.9.3
dateutils==0.6.12
python-dateutil==2.8.2
djangorestframework
mypy
pillow
model_bakery
django-filter
flake8
drf_spectacular
